Use a LabKey folder as a board

Pin data to a folder on a LabKey server

Usage

board_labkey(
  base_url,
  folder,
  subdir = "pins",
  versioned = TRUE,
  api_key = Sys.getenv("LABKEY_API_KEY"),
  cache_alias = NULL
)

Arguments

base_url

The baseUrl of the Labkey server

folder

The folder path (aka folderPath) within the LabKey server to read/write pins

subdir

The subdirectory within the LabKey folder (aka remoteFilePath) where pin should be stored (default "pins")

versioned

Boolean; whether to version the pin (default TRUE)

api_key

API key to use for LabKey authentication. If not specified, will use LABKEY_API_KEY

cache_alias

Alias of the board to use in cache (if not specified will use default pins cache location)

Value

A board object of class "pins_board_labkey"

Details

board_labkey() is powered by the Rlabkey package https://github.com/cran/Rlabkey

Examples

if (FALSE) { # \dontrun{
board <- board_labkey(
  base_url = "https://learn.labkey.com/",
  folder = "LabKey_Board/"
)
board %>% pin_write(mtcars)
board %>% pin_read("mtcars")
} # }